code/DeSCo/requirements.txt
absl-py==1.0.0
aiohttp==3.8.4
aiosignal==1.3.1
ansiterm==1.1
async-timeout==4.0.2
attrs==23.1.0
cachetools==5.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cmake==3.26.3
contourpy==1.0.7
cycler==0.11.0
deepsnap==0.2.1
filelock==3.12.0
fonttools==4.39.4
frozenlist==1.3.3
fsspec==2023.5.0
google-auth==2.19.0
google-auth-oauthlib==1.0.0
graph-tools==1.11
grpcio==1.54.2
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
lit==16.0.5
littleutils==0.2.2
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.3.0
multidict==6.0.4
munkres==1.1.4
networkx==2.7.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
ogb==1.3.6
orca==1.0
outdated==0.2.2
packaging==23.1
pandas==1.4.1
perlcompat==1.1
Pillow==10.0.1
platformdirs==3.5.1
pooch==1.7.0
protobuf==3.20.1
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycairo==1.23.0
pycparser==2.21
pyDeprecate==0.3.2
PyGObject==3.44.1
pyparsing==3.0.9
PySocks==1.7.1
Pytess==1.0.0
python-dateutil==2.8.2
pytorch-lightning==1.6.4
pytz==2023.3
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.11.2
six==1.16.0
sympy==1.12
tbdump==1.4
tensorboard==2.13.0
tensorboard-data-server==0.7.0
threadpoolctl==3.1.0
# torch==1.13.0
# torch-cluster==1.6.0
torch-geometric==2.2.0
# torch-scatter==2.0.9
# torch-sparse==0.6.15
# torch-spline-conv==1.2.1
torch-tb-profiler==0.4.0
# torchaudio==0.13.0
torchmetrics==0.11.4
# torchvision==0.14.0+cu117
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.6.2
unicodedata2==15.0.0
urllib3==1.26.16
Werkzeug==2.3.4
yarl==1.9.2
zipp==3.15.0
zstandard==0.19.0
